docs: documented how string interning works

+++ crates/common/src/symbols.rs
@@ -3,11 +3,11 @@ use std::{hash::Hash, marker::PhantomData};
 use rart::{AdaptiveRadixTree, VectorKey};
 ///A pointer to some intern string. This is 48bits for the actual position of the string in the internalized string, and 16bits for it's length
 #[derive(Debug)]
-pub struct SymbolPointer<T>(usize, PhantomData<T>);
+pub struct SymbolPointer<T>(u64, PhantomData<T>);
 
 impl<T> SymbolPointer<T> {
-    pub fn new(ptr: usize, length: u16) -> Self {
-        Self(ptr << 16 | length as usize, PhantomData)
+    pub fn new(ptr: u64, length: u16) -> Self {
+        Self(ptr << 16 | length as u64, PhantomData)
     }
 }
 
@@ -42,7 +42,7 @@ pub struct InternalizedString {
 impl InternalizedString {
     ///Inserts the provided `s` string and returns it's pointer. Note that it's size cannot be >65535
     pub fn insert<T>(&mut self, s: &str) -> SymbolPointer<T> {
-        let ptr = self.inner.len();
+        let ptr = self.inner.len() as u64;
         let size = s.len();
         self.inner.extend_from_slice(s.as_bytes());
         SymbolPointer::new(ptr, size as u16)
@@ -64,8 +64,8 @@ impl<Ctx> std::default::Default for SymbolsModule<Ctx> {
 impl<Ctx> std::ops::Index<SymbolPointer<Ctx>> for SymbolsModule<Ctx> {
     type Output = str;
     fn index(&self, ptr: SymbolPointer<Ctx>) -> &Self::Output {
-        let size = ptr.0 & 0xffff;
-        let ptr = ptr.0 >> 16;
+        let size = ptr.0 as usize & 0xffff;
+        let ptr = ptr.0 as usize >> 16;
         unsafe { std::str::from_utf8_unchecked(&self.names.inner[ptr..ptr + size]) }
     }
 }
